DB추가 및 이동 동기화 기초작업
MySQL과 Redis 서버에 추가.
클라이언트와 서버의 위치 동기화 작업

+++ MMORPG/Server/ChatPacketHandler.cpp
@@ -1,5 +1,4 @@
 #include "ChatPacketHandler.h"
-#include "IOCP.h"
 #include "MainServer.h" 
 
 ChatPacketHandler::ChatPacketHandler(MapManager& mapManager)
@@ -7,7 +6,7 @@ ChatPacketHandler::ChatPacketHandler(MapManager& mapManager)
 {
 }
 
-bool ChatPacketHandler::CanHandle(int packetID) const
+bool ChatPacketHandler::CanHandle(uint16_t packetID) const
 {
     return packetID == C2SPlayerChat;
 }
